==> Scripts/Python_Scripts/04_ML_Sales_Prediction.ipynb <==
import pandas as pd
from prophet import Prophet
Policies = pd.read_csv('../data/01_raw_data/policies.csv')

Policies.columns

# Ensure correct column names and types
Policies['StartDate'] = pd.to_datetime(Policies['StartDate'], errors='coerce')

# Now group by month using the correct column
monthly_sales = Policies.groupby(pd.Grouper(key='StartDate', freq='M'))['PremiumAmount'].sum().reset_index()

monthly_sales.rename(columns={'StartDate':'ds','PremiumAmount':'y'}, inplace = True)


monthly_sales.head()

from prophet import Prophet

model = Prophet()
model.fit(monthly_sales)

future = model.make_future_dataframe(periods=6,freq = 'M')
forecast = model.predict(future)

forecast_df = forecast[['ds', 'yhat', 'yhat_lower', 'yhat_upper']]
forecast_df.rename(columns={'ds': 'Date', 'yhat': 'Predicted', 'yhat_lower': 'LowerBound', 'yhat_upper': 'UpperBound'}, inplace=True)

# Merge actual + forecast for comparison
final_df = pd.merge(monthly_sales, forecast_df, how='right', left_on='ds', right_on='Date')


final_df.to_csv('../data/01_raw_data/sales_forecast.csv', index=False)


final_df.head(10)
